requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
EditorConfig==0.12.4
Faker==25.3.0
fast-query-parsers==1.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsbeautifier==1.15.1
litestar==2.8.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
numpy==1.26.4
openai==1.30.5
pandas==2.2.2
polyfactory==2.16.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
rich-click==1.8.2
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
